Ability to run commands from the main process
